renamed test util classes to prevent ant from executing them
