+++ src-test/magellan/library/utils/AtlantisOrderWriterTest.java
@@ -36,7 +36,7 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class AtlantisOrderWriterTest extends OrderWriterTest {
+public class AtlantisOrderWriterTest extends OrderWriterTestUtil {
 
   @BeforeClass
   public static void setUpBeforeClass() throws Exception {
